Added API integration with Shalazam to auto post nodes to the map. (requires special access)
Added Mob tracking to a separate CSV file.

Added a config file, it will be generated on first run and will save all settings in the gui. BEWARE: Your API key is saved in plain text in this file.

Settings Tab allows you to toggle features as well as set file locations and your API key.

Add a README to help with setup.